Alan-Quintanilla-habitos/microservicios/servicio_a/requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
requests==2.31.0
